Environment API, part 2
In this part we simply re-arrange source files
