+++ core/mcp.go
@@ -48,61 +48,6 @@ type MCP struct {
 	functionMask map[string]bool
 }
 
-type Environment struct {
-	// Saved objects by prompt var name
-	objsByName map[string]*Binding
-	// Saved objects by ID (Foo#123)
-	objsByID map[string]*Binding
-	// String variables assigned to the environment
-	varsByName map[string]*Binding
-	// Auto incrementing number per-type
-	typeCount map[string]int
-	// The LLM-friendly ID ("Container#123") for each object
-	idByHash map[digest.Digest]string
-}
-
-type Binding struct {
-	Key   string
-	Value dagql.Typed
-	env   *Environment // TODO: wire this up
-}
-
-func (b *Binding) AsObject() (dagql.Object, bool) {
-	obj, ok := dagql.UnwrapAs[dagql.Object](b.Value)
-	return obj, ok
-}
-
-func (b *Binding) AsEnumerable() (dagql.Enumerable, bool) {
-	enum, ok := dagql.UnwrapAs[dagql.Enumerable](b.Value)
-	return enum, ok
-}
-
-func (b *Binding) TypeName() string {
-	if b.Value == nil {
-		return Void{}.TypeName()
-	}
-	return b.Value.Type().Name()
-}
-
-// Return the stable object ID for this binding, or an empty string if it's not an object
-func (b *Binding) ID() string {
-	obj, isObject := b.AsObject()
-	if !isObject {
-		return ""
-	}
-	return b.env.Ingest(obj)
-}
-
-func NewEnvironment() *Environment {
-	return &Environment{
-		objsByName: map[string]*Binding{},
-		objsByID:   map[string]*Binding{},
-		varsByName: map[string]*Binding{},
-		typeCount:  map[string]int{},
-		idByHash:   map[digest.Digest]string{},
-	}
-}
-
 func NewMCP(endpoint *LLMEndpoint) *MCP {
 	return &MCP{
 		env:               NewEnvironment(),
@@ -128,16 +73,6 @@ func (m *MCP) Clone() *MCP {
 	return &cp
 }
 
-func (env *Environment) Clone() *Environment {
-	cp := *env
-	cp.objsByName = cloneMap(cp.objsByName)
-	cp.objsByID = cloneMap(cp.objsByID)
-	cp.varsByName = cloneMap(cp.varsByName)
-	cp.typeCount = cloneMap(cp.typeCount)
-	cp.idByHash = cloneMap(cp.idByHash)
-	return &cp
-}
-
 // Return the current selection
 func (m *MCP) Current() dagql.Object {
 	return m.current
@@ -157,61 +92,6 @@ func (m *MCP) Set(key string, obj dagql.Object) {
 	m.env = m.env.WithBinding(key, obj)
 }
 
-// Add a binding to the environment
-func (env *Environment) WithBinding(key string, obj dagql.Object) *Environment {
-	env = env.Clone()
-	env.objsByName[key] = &Binding{Key: key, Value: obj}
-	env.Ingest(obj)
-	return env
-}
-
-// List all object bindings in the environment
-// TODO: expand from "object bindings" to "all bindings"
-func (env *Environment) Bindings() []*Binding {
-	res := make([]*Binding, 0, len(env.objsByName))
-	for _, v := range env.objsByName {
-		res = append(res, v)
-	}
-	return res
-}
-
-// Return all object bindings of the given type
-// TODO: expand beyond object types
-func (env *Environment) BindingsOfType(typename string) []*Binding {
-	res := make([]*Binding, 0, len(env.objsByName))
-	for _, v := range env.objsByName {
-		if v.TypeName() == typename {
-			res = append(res, v)
-		}
-	}
-	return res
-}
-
-// Add a string variable to the environment
-// TODO: merge into bindings
-func (env *Environment) WithVariable(name, value string) *Environment {
-	env = env.Clone()
-	env.varsByName[name] = &Binding{Key: name, Value: dagql.NewString(value), env: env}
-	return env
-}
-
-// Retrieve a string variable
-// TODO: merge into bindings
-func (env *Environment) Variable(name string) (*Binding, bool) {
-	b, found := env.varsByName[name]
-	return b, found
-}
-
-// List all variables
-// TODO: merge into bindings
-func (env *Environment) Variables() []*Binding {
-	res := make([]*Binding, 0, len(env.varsByName))
-	for _, v := range env.varsByName {
-		res = append(res, v)
-	}
-	return res
-}
-
 // Get an object saved at a given key
 func (m *MCP) GetObject(key, expectedType string) (dagql.Object, error) {
 	if expectedType != "" {
@@ -229,27 +109,9 @@ func (m *MCP) GetObject(key, expectedType string) (dagql.Object, error) {
 	return nil, fmt.Errorf("unknown object %q", key)
 }
 
-func (env *Environment) Binding(key string) (*Binding, bool) {
-	// next check for values by ID
-	if val, exists := env.objsByID[key]; exists {
-		return val, true
-	}
-	// next check for values by name
-	if val, exists := env.objsByName[key]; exists {
-		return val, true
-	}
-	return nil, false
-}
-
 // Unset a saved value
 func (m *MCP) Unset(key string) {
 	m.env = m.env.WithoutBinding(key)
-}
-
-func (env *Environment) WithoutBinding(key string) *Environment {
-	env = env.Clone()
-	delete(env.objsByName, key)
-	return env
 }
 
 func (m *MCP) Tools(srv *dagql.Server) ([]LLMTool, error) {
@@ -540,36 +402,6 @@ func (m *MCP) describe(val dagql.Typed) string {
 	return val.Type().String()
 }
 
-// Return a stable digest of the binding's value
-func (b *Binding) Digest() digest.Digest {
-	obj, isObject := b.AsObject()
-	if isObject {
-		return obj.ID().Digest()
-	}
-	jsonBytes, err := json.Marshal(b.Value)
-	if err != nil {
-		return digest.FromString("")
-	}
-	return dagql.HashFrom(string(jsonBytes))
-}
-
-func (env *Environment) Ingest(obj dagql.Object) string {
-	id := obj.ID()
-	if id == nil {
-		return ""
-	}
-	hash := id.Digest()
-	typeName := id.Type().NamedType()
-	llmID, ok := env.idByHash[hash]
-	if !ok {
-		env.typeCount[typeName]++
-		llmID = fmt.Sprintf("%s#%d", typeName, env.typeCount[typeName])
-		env.idByHash[hash] = llmID
-		env.objsByID[llmID] = &Binding{Key: llmID, Value: obj, env: env}
-	}
-	return llmID
-}
-
 func (m *MCP) Call(ctx context.Context, tools []LLMTool, toolCall ToolCall) (string, bool) {
 	var tool *LLMTool
 	for _, t := range tools {
@@ -657,22 +489,6 @@ func (m *MCP) ReadVariable(name string) (string, bool) {
 		return v.AsString()
 	}
 	return "", false
-}
-
-func (b *Binding) AsString() (string, bool) {
-	s, ok := dagql.UnwrapAs[dagql.String](b.Value)
-	if !ok {
-		return "", false
-	}
-	return s.String(), true
-}
-
-func (env *Environment) Types() []string {
-	types := make([]string, 0, len(env.typeCount))
-	for typ := range env.typeCount {
-		types = append(types, typ)
-	}
-	return types
 }
 
 func (m *MCP) Builtins(srv *dagql.Server) ([]LLMTool, error) {
